make msh more efficient
- changed msh_get_Var to msh_var_get..., msh_push_Var to msh_var_push
- also implemeted msh_var_update..., like with the local vars

This change aims to copy less and maybe unify the var getter and setters
for local and global in the future.
This should make msh more efficient and less likely to have a
stack-overflow, which I think, happens, because recently I get
random segmentation faults, only on MacOS and not while debugging...

Next step will be to make a msh_var_copy or similar
and use it instead of word_copy, in order to limit the copying
to VAR_MAXCHAR (4000 bytes).
I'm also thinking to put VAR_SPEICHER and maybe also
VAR_NAMES in the heap...